1.添加库位模块

+++ src/main/java/com/gz/xg/domain/vo/ProdTagVo.java
@@ -24,7 +24,7 @@ public class ProdTagVo {
 
     private String productCategory;
 
-    private BigDecimal qty;
+    private Integer qty;
 
     private String prodNo;
